Write a C program that prints the size of various types on the computer it is compiled and run on, fix

+++ 0x00-hello_world/6-size.c
@@ -10,7 +10,7 @@ int main(void)
 	printf("Size of char is %ld byte(s)\n", sizeof(char));
 	printf("Size of int is %ld byte(s)\n", sizeof(int));
 	printf("Size of long int is %ld byte(s)\n", sizeof(long));
-	printf("Size of long long int is %lld byte(s)\n", sizeof(long long int));
+	printf("Size of long long int is %lld byte(s)\n", sizeof(long));
 	printf("Size of float is %ld byte(s)\n", sizeof(float));
 	return (0);
 }
